added latte model tests

+++ src/database/models.py
@@ -64,4 +64,4 @@ class Latte(db.Model):
         db.session.commit()
 
     def __repr__(self):
-        return json.dumps(self.short())
+        return json.dumps(self.long())
